Add API based hydration and field extension
Requests to pgstac search functions can include a conf setting to
prevent the database from hydrating search results. Adds a setting to
stac-fastapi instances that applies the conf setting and hydrates and
filters fields within the API. This is useful when the DB hydration
presents performance bottlenecks due to high cpu usage, which could
otherwise be spread to API servers.